requirements.txt
beautifulsoup4==4.12.3
pandas==2.2.2
python-dotenv==1.0.1
requests==2.32.3
openpyxl==3.1.5